backend/requirements.txt
Django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
django-filter
python-decouple
drf-spectacular
Pillow
psycopg2-binary
gunicorn
# dj-database-url
dj-database-url
